open_node: introduce jlink base node and use it for nrf52dk
