+++ gateway_code/static/iot-lab-nrf52xxxdk.cfg
